corrected the little bug
There was a little bug on every double click, because the section where the program checks if the current box is empty or not, the checking part semicolon was not covered the whole part. That is why was allowed to 'double click' to an element. Now even if I double click doesn't generate more item, just the one prober index value.